pyproject.toml
[project]
name = "health-bay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.18.0",
    "python-dotenv>=1.1.0",
]
